requirements.txt
discord.py>=2.4.0
pymongo
python-dotenv
Levenshtein
pillow
numpy